+++ src/main/java/com/gamesense/client/module/ModuleManager.java
@@ -39,8 +39,7 @@ public class ModuleManager {
 		addMod(new KillAura());
 		addMod(new OffhandCrystal());
 		addMod(new OffhandGap());
-		//addMod(new PistonCrystal());
-		addMod(new PistonCrystalRewritten());
+		addMod(new PistonCrystal());
 		addMod(new SelfTrap());
 		addMod(new SelfWeb());
 		addMod(new Surround());
