--- src/main/java/com/gamesense/client/module/modules/combat/PistonCrystalRewritten.java
@@ -1,1367 +1,0 @@
-package com.gamesense.client.module.modules.combat;
-
-/**
- * @Author TechAle on (Insert data official release)
- * Ported and modified from AutoAnvil.java that is modified from Surround.java
- * Break crystal from AutoCrystal
- */
-
-
-import com.gamesense.api.event.events.PacketEvent;
-import com.gamesense.api.setting.Setting;
-import com.gamesense.api.util.misc.MessageBus;
-import com.gamesense.api.util.player.friends.Friends;
-import com.gamesense.api.util.world.BlockUtil;
-import com.gamesense.client.module.Module;
-import com.gamesense.client.module.ModuleManager;
-import com.gamesense.client.module.modules.gui.ColorMain;
-import me.zero.alpine.listener.EventHandler;
-import me.zero.alpine.listener.Listener;
-import net.minecraft.block.*;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.item.EntityEnderCrystal;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.*;
-import net.minecraft.network.Packet;
-import net.minecraft.network.play.client.CPacketEntityAction;
-import net.minecraft.network.play.client.CPacketPlayer;
-import net.minecraft.network.play.client.CPacketPlayerDigging;
-import net.minecraft.network.play.client.CPacketUseEntity;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class PistonCrystalRewritten extends Module {
-
-    public PistonCrystalRewritten() {
-        super("PistonCrystalRewritten", Category.Combat);
-    }
-
-    Setting.Mode    breakType,
-                    placeMode,
-                    target;
-
-    Setting.Double  enemyRange,
-                    torchRange,
-                    crystalDeltaBreak;
-
-    Setting.Integer blocksPerTick,
-                    startDelay,
-                    supBlocksDelay,
-                    pistonDelay,
-                    crystalDelay,
-                    hitDelay,
-                    midHitDelay,
-                    stuckDetector,
-                    maxYincr;
-    Setting.Boolean rotate,
-                    blockPlayer,
-                    confirmBreak,
-                    confirmPlace,
-                    allowFastMode,
-                    betterPlacement,
-                    bypassObsidian,
-                    antiWeakness,
-                    debugMode,
-                    chatMsg;
-
-    private boolean noMaterials = false,
-                    hasMoved = false,
-                    isSneaking = false,
-                    yUnder = false,
-                    isHole = true,
-                    enoughSpace = true,
-                    redstoneBlockMode = false,
-                    fastModeActive = false,
-                    broken,
-                    brokenCrystalBug,
-                    brokenRedstoneTorch,
-                    stoppedCa,
-                    deadPl;
-
-    private int oldSlot = -1,
-            stage,
-            delayTimeTicks,
-            stuck = 0,
-            hitTryTick;
-
-    private int[]   slot_mat,
-                    delayTable,
-            meCoordsInt,
-                    enemyCoordsInt;
-
-    private double[] enemyCoordsDouble;
-
-    private structureTemp toPlace;
-
-    int[][] disp_surblock = {
-            {1,0,0},
-            {-1,0,0},
-            {0,0,1},
-            {0,0,-1}
-    };
-
-    Double[][] sur_block = new Double[4][3];
-
-    private EntityPlayer aimTarget;
-
-
-    public void setup(){
-        ArrayList<String> breakTypes = new ArrayList<>();
-        breakTypes.add("Swing");
-        breakTypes.add("Packet");
-        ArrayList<String> placeModes = new ArrayList<>();
-        placeModes.add("Torch");
-        placeModes.add("Block");
-        placeModes.add("Both");
-        ArrayList<String> targetChoose = new ArrayList<>();
-        targetChoose.add("Nearest");
-        targetChoose.add("Looking");
-        target = registerMode("Target", "Target", targetChoose, "Nearest");
-        breakType = registerMode("Type", "Type", breakTypes, "Swing");
-        placeMode = registerMode("Place", "Place", placeModes, "Torch");
-        enemyRange = registerDouble("Range", "Range",4.9, 0, 6);
-        torchRange = registerDouble("TorchRange", "TorchRange",5.5, 0, 6);
-        crystalDeltaBreak = registerDouble("crystalDeltaBreak", "crystalDeltaBreak",0.1, 0, 0.5);
-        blocksPerTick = registerInteger("Blocks Per Tick", "BlocksPerTick", 4, 0, 20);
-        startDelay = registerInteger("Start Delay", "StartDelay", 4, 0, 20);
-        stuckDetector = registerInteger("Stuck Check", "StuckCheck", 35, 0, 200);
-        pistonDelay = registerInteger("Piston Delay", "PistonDelay", 2, 0, 20);
-        crystalDelay = registerInteger("Crystal Delay", "Crystal Delay", 2, 0, 20);
-        hitDelay = registerInteger("Hit Delay", "HitDelay", 2, 0, 20);
-        midHitDelay = registerInteger("midHitDelay", "midHitDelay", 5, 0, 20);
-        supBlocksDelay = registerInteger("Surround Delay", "SurroundDelay", 4, 0, 20);
-        maxYincr = registerInteger("Max Y", "MaxY", 3, 0, 5);
-        rotate = registerBoolean("Rotate", "Rotate", false);
-        blockPlayer = registerBoolean("Trap Player", "TrapPlayer", true);
-        bypassObsidian = registerBoolean("BypassObsidian", "BypassObsidian", false);
-        confirmBreak = registerBoolean("No Glitch Break", "NoGlitchBreak", true);
-        confirmPlace = registerBoolean("No Glitch Place", "NoGlitchPlace", true);
-        allowFastMode = registerBoolean("Allow Fast Mode", "allowFastMode", false);
-        betterPlacement = registerBoolean("Better Place", "betterPlacement", true);
-        antiWeakness = registerBoolean("Anti Weakness", "AntiWeakness", false);
-        debugMode = registerBoolean("debugMode", "debugMode", false);
-        chatMsg = registerBoolean("Chat Msgs", "ChatMsgs", true);
-    }
-
-    public void onEnable() {
-        // Init values
-        initValues();
-        // Get Target
-        if (getAimTarget())
-            return;
-        // Make checks
-        playerChecks();
-
-    }
-
-    // Get target function
-    private boolean getAimTarget() {
-        /// Get aimTarget
-        // If nearest, get it
-        if (target.getValue().equals("Nearest"))
-            aimTarget = findClosestTarget(enemyRange.getValue(), aimTarget);
-        // if looking
-        else
-            aimTarget = findLookingPlayer(enemyRange.getValue());
-
-        // If we didnt found a target
-        if (aimTarget == null || !target.getValue().equals("Looking")){
-            // if it's not looking and we didnt found a target
-            if (!target.getValue().equals("Looking") && aimTarget == null)
-                disable();
-            // If not found a target
-            if (aimTarget == null)
-                return true;
-        }
-        return false;
-    }
-
-    // Make some checks for startup
-    private void playerChecks() {
-        // Get all the materials
-        if (getMaterialsSlot()) {
-            // check if the enemy is in a hole
-            if (is_in_hole()) {
-                // Get enemy coordinates
-                enemyCoordsDouble = new double[] {aimTarget.posX, aimTarget.posY, aimTarget.posZ};
-                enemyCoordsInt = new int[] {(int) enemyCoordsDouble[0], (int) enemyCoordsDouble[1], (int) enemyCoordsDouble[2]};
-                // Get me coordinates
-                meCoordsInt = new int[] {(int) mc.player.posX, (int) mc.player.posY, (int) mc.player.posZ};
-                // Start choosing where to place what
-                enoughSpace = createStructure();
-            // Is not in a hoke
-            } else {
-                isHole = false;
-            }
-        // No materials
-        }else noMaterials = true;
-    }
-
-    // Init some values
-    private void initValues() {
-        // Reset aimtarget
-        aimTarget = null;
-        // Create new delay table
-        delayTable = new int[] {
-                startDelay.getValue(),
-                supBlocksDelay.getValue(),
-                pistonDelay.getValue(),
-                crystalDelay.getValue(),
-                hitDelay.getValue()
-        };
-        // Default values reset
-        toPlace = new structureTemp(0,0,null);
-        isHole = true;
-        hasMoved = deadPl = broken = brokenCrystalBug = brokenRedstoneTorch = yUnder = redstoneBlockMode = fastModeActive = false;
-        slot_mat = new int[]{-1, -1, -1, -1, -1, -1};
-        stage = delayTimeTicks = stuck = 0;
-
-        if (mc.player == null){
-            disable();
-            return;
-        }
-
-        if (chatMsg.getValue()){
-            printChat("PistonCrystal turned ON!", false);
-        }
-
-        oldSlot = mc.player.inventory.currentItem;
-        // Disable ca
-        // Stop CA
-        stoppedCa = false;
-
-        if (ModuleManager.isModuleEnabled("AutoCrystalGS")){
-            AutoCrystal.stopAC = true;
-            stoppedCa = true;
-        }
-    }
-
-    // On disable of the module
-    public void onDisable() {
-        if (mc.player == null){
-            return;
-        }
-        // If output
-        if (chatMsg.getValue()){
-            String output = "";
-            // No target found
-            if (aimTarget == null) {
-                output = "No target found...";
-            }else
-            // H distance not avaible
-            if (yUnder) {
-                output = "Sorry but you cannot be 2+ blocks under the enemy or %d above...";
-            // No Materials
-            }else if (noMaterials){
-                output = "No Materials Detected...";
-            // No Hole
-            }else if (!isHole) {
-                output = "The enemy is not in a hole...";
-            // No Space
-            }else if(!enoughSpace) {
-                output = "Not enough space...";
-            // Has Moved
-            }else if(hasMoved) {
-               output = "Out of range...";
-            }else if(deadPl) {
-                output = "Enemy is dead, gg! ";
-            }
-            // Output in chat
-            printChat(output + "PistonCrystal turned OFF!", true);
-
-            // Re-Active ca
-            if (stoppedCa){
-                AutoCrystal.stopAC = false;
-                stoppedCa = false;
-            }
-
-        }
-
-        if (isSneaking){
-            mc.player.connection.sendPacket(new CPacketEntityAction(mc.player, CPacketEntityAction.Action.STOP_SNEAKING));
-            isSneaking = false;
-        }
-
-        if (oldSlot != mc.player.inventory.currentItem && oldSlot != -1){
-            mc.player.inventory.currentItem = oldSlot;
-            oldSlot = -1;
-        }
-
-        noMaterials = false;
-        AutoCrystal.stopAC = false;
-    }
-
-    // Every updates
-    public void onUpdate() {
-        // If no mc.player
-        if (mc.player == null){
-            disable();
-            return;
-        }
-
-        // Wait
-        if (delayTimeTicks < delayTable[stage]){
-            delayTimeTicks++;
-            return;
-        }
-        // If the delay is finished
-        else {
-            delayTimeTicks = 0;
-        }
-
-        // Check if something is not ok
-        if (enemyCoordsDouble == null || aimTarget == null) {
-            if (aimTarget == null) {
-                aimTarget = findLookingPlayer(enemyRange.getValue());
-                if (aimTarget != null) {
-                    playerChecks();
-                }
-            }else
-                checkVariable();
-            return;
-        }
-
-        // Check if he is dead
-        if (aimTarget.isDead) {
-            deadPl = true;
-        }
-
-        // Check if he is not in the hole
-        if ((int) aimTarget.posX != (int) enemyCoordsDouble[0] || (int) aimTarget.posZ != (int) enemyCoordsDouble[2])
-            hasMoved = true;
-
-        // If we have to left
-        if (checkVariable()){
-            return;
-        }
-
-        /*
-            This is how it works:
-            a)
-                First of all, it check if  we have all the supportive blocks + if we have arleady something.
-                (i want it to prevent some bugs)
-            If there is some missing one, it is going to place a block and it will finish.
-            then, it place:
-                b) Piston
-                c) Crystal
-                d) redstone torch
-            then, it wait and
-                e) break the crystal.
-         */
-
-        /// Start Placing ///
-        // A) Lets place all the supports blocks
-        if (placeSupport()) {
-            switch (stage) {
-                case 1:
-                    if (breakRedstone())
-                        placeBlockThings(stage);
-                    break;
-
-                case 2:
-                    // Check pistonPlace if confirmPlace
-                    if (fastModeActive || !confirmPlace.getValue() || checkPistonPlace())
-                        placeBlockThings(stage);
-                    break;
-
-                case 3:
-                    // Check crystal if confirmPlace
-                    if (fastModeActive || !confirmPlace.getValue() || checkCrystalPlace()) {
-                        placeBlockThings(stage);
-                        hitTryTick = 0;
-                    }
-                    break;
-                case 4:
-                    destroyCrystalAlgo();
-                    break;
-            }
-        }
-
-    }
-
-    public void destroyCrystalAlgo() {
-        // Get the crystal
-        Entity crystal = null;
-        // Check if the crystal exist
-        for(Entity t : mc.world.loadedEntityList) {
-            // If it's a crystal
-            if (t instanceof EntityEnderCrystal) {
-                // Check if the crystal is in the enemy
-                if (    (int) t.posX == enemyCoordsInt[0] &&
-                    ( (int) (t.posZ - crystalDeltaBreak.getValue()) == enemyCoordsInt[2] || (int) ( t.posZ + crystalDeltaBreak.getValue()) == enemyCoordsInt[2])
-                    ||  (int) t.posZ == enemyCoordsInt[2] &&
-                    ( (int) ( t.posX - crystalDeltaBreak.getValue()) == enemyCoordsInt[0] || (int) ( t.posX + crystalDeltaBreak.getValue()) == enemyCoordsInt[0]))
-                    // If found, yoink
-                    crystal = t;
-            }
-        }
-        // If we have confirmBreak, we have found 0 crystal and we broke a crystal before
-        if (confirmBreak.getValue() && broken && crystal == null) {
-            /// That means the crystal was broken 100%
-            // Reset
-            stage = stuck = 0;
-            broken = false;
-        }
-        // If found the crystal
-        if (crystal != null) {
-            // Break it
-            breakCrystalPiston(crystal);
-            // If we have to check
-            if (confirmBreak.getValue())
-                broken = true;
-                // If not, left
-            else
-                stage = stuck = 0;
-        }else {
-            // If it got stuck
-            if (++stuck >= stuckDetector.getValue()) {
-                /// Try to find the error
-                // First error: crystal not found
-                boolean found = false;
-                for(Entity t : mc.world.loadedEntityList) {
-                    // If crystal
-                    if (t instanceof EntityEnderCrystal
-                            // If coordinates the same as where is the crystal
-                            && (int) t.posX == (int) toPlace.to_place.get(toPlace.supportBlock + 1).x &&
-                            (int) t.posZ == (int) toPlace.to_place.get(toPlace.supportBlock + 1).z ) {
-                        // Found
-                        found = true;
-                        break;
-                    }
-                }
-                // If not
-                if (!found) {
-                    //// The error is that the crystal has not been placed
-                    /// Destroy the redstone torch
-                    // If rotation
-                    BlockPos offsetPosPist = new BlockPos(toPlace.to_place.get(toPlace.supportBlock + 2));
-                    BlockPos pos = new BlockPos(aimTarget.getPositionVector()).add(offsetPosPist.getX(), offsetPosPist.getY(), offsetPosPist.getZ());
-
-                    // Check if there is the redstone torch. It is has been destroyed before
-                    if (confirmBreak.getValue() && brokenRedstoneTorch && get_block(pos.x, pos.y, pos.z) instanceof BlockAir) {
-                        // Reset
-                        stage = 1;
-                        brokenRedstoneTorch = false;
-                    } else {
-                        // Else, get the side
-                        EnumFacing side = BlockUtil.getPlaceableSide(pos);
-                        if (side != null) {
-                            // If rotate, look at the redstone torch
-                            if (rotate.getValue()) {
-                                BlockPos neighbour = pos.offset(side);
-                                EnumFacing opposite = side.getOpposite();
-                                Vec3d hitVec = new Vec3d(neighbour).add(0.5, 1, 0.5).add(new Vec3d(opposite.getDirectionVec()).scale(0.5));
-                                BlockUtil.faceVectorPacketInstant(hitVec);
-
-                            }
-                            // Destroy it
-                            mc.player.swingArm(EnumHand.MAIN_HAND);
-                            mc.player.connection.sendPacket(new CPacketPlayerDigging(
-                                    CPacketPlayerDigging.Action.START_DESTROY_BLOCK, pos, side
-                            ));
-                            mc.player.connection.sendPacket(new CPacketPlayerDigging(
-                                    CPacketPlayerDigging.Action.STOP_DESTROY_BLOCK, pos, side
-                            ));
-                            /// Restart from the crystal
-                            if (confirmBreak.getValue())
-                                brokenRedstoneTorch = true;
-                            else
-                                stage = 1;
-                            // print
-                            printChat("Stuck detected: crystal not placed", true);
-                        }
-                    }
-
-                }else {
-                    // Try to see if the crystal, somehow, is in the piston extended thing (it happened sometimes)
-                    boolean ext = false;
-                    // Find the crystal
-                    for(Entity t : mc.world.loadedEntityList) {
-                        if (t instanceof EntityEnderCrystal
-                                && (int) t.posX == (int) toPlace.to_place.get(toPlace.supportBlock + 1).x &&
-                                (int) t.posZ == (int) toPlace.to_place.get(toPlace.supportBlock + 1).z ) {
-                            ext = true;
-                            break;
-                        }
-                    }
-                    // If it hasnt been found and confirmBreak
-                    if (confirmBreak.getValue() && brokenCrystalBug && !ext) {
-                        // Reset
-                        stage = stuck = 0;
-                        brokenCrystalBug = false;
-                    }
-                    // If yes
-                    if (ext) {
-                        // Break the crystal
-                        breakCrystalPiston(crystal);
-                        // If we have to check
-                        if (confirmBreak.getValue())
-                            brokenCrystalBug = true;
-                        else
-                            stage = stuck = 0;
-                        // Print
-                        printChat("Stuck detected: crystal is stuck in the moving piston", true);
-                    }
-                }
-            }
-        }
-    }
-
-    private void breakCrystalPiston (Entity crystal) {
-        // HitDelay
-        if (hitTryTick++ < midHitDelay.getValue())
-            return;
-        else
-            hitTryTick = 0;
-        // If weaknes
-        if (antiWeakness.getValue())
-            mc.player.inventory.currentItem = slot_mat[4];
-        // If rotate
-        if (rotate.getValue()) {
-            lookAtPacket(crystal.posX, crystal.posY, crystal.posZ, mc.player);
-        }
-        /// Break type
-        // Swing
-        if (breakType.getValue().equals("Swing")) {
-            breakCrystal(crystal);
-            // Packet
-        }else if(breakType.getValue().equals("Packet")) {
-            mc.player.connection.sendPacket(new CPacketUseEntity(crystal));
-            mc.player.swingArm(EnumHand.MAIN_HAND);
-        }
-        // Rotate
-        if (rotate.getValue())
-            resetRotation();
-    }
-
-    private boolean breakRedstone() {
-        BlockPos offsetPosPist = new BlockPos(toPlace.to_place.get(toPlace.supportBlock + 2));
-        BlockPos pos = new BlockPos(aimTarget.getPositionVector()).add(offsetPosPist.getX(), offsetPosPist.getY(), offsetPosPist.getZ());
-        if (!(get_block(pos.x, pos.y , pos.z) instanceof BlockAir)) {
-            breakBlock(pos);
-            return false;
-        }
-        return true;
-    }
-
-    private void breakBlock(BlockPos pos) {
-        // If we have a redstone block
-        if (redstoneBlockMode) {
-            // Switch to the pick
-            mc.player.inventory.currentItem = slot_mat[5];
-        }
-        EnumFacing side = BlockUtil.getPlaceableSide(pos);
-        if (side != null) {
-            // If rotate, look at the redstone torch
-            if (rotate.getValue()) {
-                BlockPos neighbour = pos.offset(side);
-                EnumFacing opposite = side.getOpposite();
-                Vec3d hitVec = new Vec3d(neighbour).add(0.5, 1, 0.5).add(new Vec3d(opposite.getDirectionVec()).scale(0.5));
-                BlockUtil.faceVectorPacketInstant(hitVec);
-
-            }
-            // Destroy it
-            mc.player.swingArm(EnumHand.MAIN_HAND);
-            mc.player.connection.sendPacket(new CPacketPlayerDigging(
-                    CPacketPlayerDigging.Action.START_DESTROY_BLOCK, pos, side
-            ));
-            mc.player.connection.sendPacket(new CPacketPlayerDigging(
-                    CPacketPlayerDigging.Action.STOP_DESTROY_BLOCK, pos, side
-            ));
-        }
-    }
-
-    private boolean checkPistonPlace() {
-        // Check for the piston
-        BlockPos targetPosPist = compactBlockPos(stage - 1);
-        if (!(get_block(targetPosPist.x, targetPosPist.y, targetPosPist.z) instanceof BlockPistonBase)) {
-            // Go back placing the piston
-            stage--;
-            return false;
-        }else return true;
-    }
-
-    private boolean checkCrystalPlace() {
-        // Check if the crystal has been placed
-        for (Entity t : mc.world.loadedEntityList) {
-            // If is an endcrystal
-            if (t instanceof EntityEnderCrystal
-                    // If the position is right
-                    && (int) t.posX == (int) (aimTarget.posX + toPlace.to_place.get(toPlace.supportBlock + 1).x) &&
-                    (int) t.posZ == (int) (aimTarget.posZ + toPlace.to_place.get(toPlace.supportBlock + 1).z)) {
-                return true;
-            }
-        }
-
-        stage--;
-
-        return false;
-    }
-
-    private boolean placeSupport() {
-        // N^ blocks checked
-        int checksDone = 0;
-        // N^ blocks placed
-        int blockDone = 0;
-        // If we have to place
-        if (toPlace.supportBlock > 0) {
-            // Lets iterate and check
-            // Lets finish
-            do {
-                BlockPos targetPos = getTargetPos(checksDone);
-
-                // Try to place the block
-                if (placeBlock(targetPos, 0, 0, 0, 1)) {
-                    // If we reached the limit
-                    if (++blockDone == blocksPerTick.getValue())
-                        // Return false
-                        return false;
-                }
-
-            // If we reached the limit, exit
-            } while (++checksDone != toPlace.supportBlock);
-        }
-        stage = stage == 0 ? 1 : stage;
-        return true;
-    }
-
-    private boolean placeBlock(BlockPos pos, int step, double offsetX, double offsetZ, double offsetY){
-        // Get the block
-        Block block = mc.world.getBlockState(pos).getBlock();
-        // Get all sides
-        EnumFacing side = BlockUtil.getPlaceableSide(pos);
-
-        // If there is a solid block
-        if (!(block instanceof BlockAir) && !(block instanceof BlockLiquid)){
-            return false;
-        }
-        // If we cannot find any side
-        if (side == null){
-            return false;
-        }
-
-        // Get position of the side
-        BlockPos neighbour = pos.offset(side);
-        EnumFacing opposite = side.getOpposite();
-
-
-        // If that block can be clicked
-        if (!BlockUtil.canBeClicked(neighbour)){
-            return false;
-        }
-
-        // Get the position where we are gonna click
-        Vec3d hitVec = new Vec3d(neighbour).add(0.5 + offsetX, offsetY, 0.5 + offsetZ).add(new Vec3d(opposite.getDirectionVec()).scale(0.5));
-        Block neighbourBlock = mc.world.getBlockState(neighbour).getBlock();
-
-        /*
-			// I use this as a remind to which index refers to what
-			0 => obsidian
-			1 => piston
-			2 => Crystals
-			3 => redstone
-		 */
-        // Get what slot we are going to select
-        // If it's not empty
-        if (mc.player.inventory.getStackInSlot(slot_mat[step]) != ItemStack.EMPTY) {
-            // Is it is correct
-            if (mc.player.inventory.currentItem != slot_mat[step]) {
-                // Change the hand's item
-                mc.player.inventory.currentItem = slot_mat[step] == 11 ? mc.player.inventory.currentItem : slot_mat[step];
-            }
-        }else return false;
-
-        // Why?
-        if (!isSneaking && BlockUtil.blackList.contains(neighbourBlock) || BlockUtil.shulkerList.contains(neighbourBlock)){
-            mc.player.connection.sendPacket(new CPacketEntityAction(mc.player, CPacketEntityAction.Action.START_SNEAKING));
-            isSneaking = true;
-        }
-
-        // For the rotation
-        if (rotate.getValue() || step == 1){
-            Vec3d positionHit = hitVec;
-            // if rotate and we are not in rotate
-            if (!rotate.getValue() && step == 1) {
-                // Dont look at the block but just the direction
-                positionHit = new Vec3d(mc.player.posX + offsetX, mc.player.posY + (offsetY == -1 ? offsetY : 0), mc.player.posZ + offsetZ);
-            }
-            // Look
-            BlockUtil.faceVectorPacketInstant(positionHit);
-        }
-        // If we are placing with the main hand
-        EnumHand handSwing = EnumHand.MAIN_HAND;
-        // If we are placing with the offhand
-        if (slot_mat[step] == 11)
-            handSwing = EnumHand.OFF_HAND;
-
-        // Place the block
-        mc.playerController.processRightClickBlock(mc.player, mc.world, neighbour, opposite, hitVec, handSwing);
-        mc.player.swingArm(handSwing);
-
-        return true;
-    }
-
-    public void placeBlockThings(int step) {
-        // Get absolute position
-        BlockPos targetPos = compactBlockPos(step);
-        // Place
-        placeBlock(targetPos, step, toPlace.offsetX, toPlace.offsetZ, toPlace.offsetY);
-        // Next step
-        stage++;
-    }
-
-    public BlockPos compactBlockPos(int step) {
-        // Get enemy's relative position of the block
-        BlockPos offsetPos = new BlockPos(toPlace.to_place.get(toPlace.supportBlock + step - 1));
-        // Get absolute position and return
-        return new BlockPos(enemyCoordsDouble[0] + offsetPos.getX(), enemyCoordsDouble[1] + offsetPos.getY(), enemyCoordsDouble[2] + offsetPos.getZ());
-
-    }
-
-    private BlockPos getTargetPos(int idx) {
-        BlockPos offsetPos = new BlockPos(toPlace.to_place.get(idx));
-        return new BlockPos(enemyCoordsDouble[0] + offsetPos.x, enemyCoordsDouble[1] + offsetPos.y, enemyCoordsDouble[2] + offsetPos.z);
-    }
-
-    // Check if we have to disable
-    private boolean checkVariable() {
-        if (noMaterials || !isHole || !enoughSpace || hasMoved || deadPl){
-            disable();
-            return true;
-        }
-        return false;
-    }
-
-    // Class for the structure
-    private static class structureTemp {
-        public double distance;
-        public int supportBlock;
-        public List<Vec3d> to_place;
-        public int direction;
-        public float offsetX;
-        public float offsetY;
-        public float offsetZ;
-
-        public structureTemp(double distance, int supportBlock, List<Vec3d> to_place) {
-            this.distance = distance;
-            this.supportBlock = supportBlock;
-            this.to_place = to_place;
-            this.direction = -1;
-        }
-
-        public void replaceValues(double distance, int supportBlock, List<Vec3d> to_place, int direction, float offsetX, float offsetZ, float offsetY) {
-            this.distance = distance;
-            this.supportBlock = supportBlock;
-            this.to_place = to_place;
-            this.direction = direction;
-            this.offsetX = offsetX;
-            this.offsetZ = offsetZ;
-            this.offsetY = offsetY;
-        }
-    }
-
-    // Create the skeleton of the structure
-    private boolean createStructure() {
-        /*
-            I decided to split rotate and non rotate when needed.
-            I wont do like i did in the last pistonCrystal that was all toogheter
-            (yes, it's less code, but it's more complex with all in one)
-         */
-        /// Create default start value
-        // Structure default
-        structureTemp addedStructure = new structureTemp(Double.MAX_VALUE, 0, null);
-        // Distance we are going to find
-        double distanceNowCrystal;
-        // Since they may happens some errors that i did not expect, i use a try-catch
-        try {
-            // First check, h check.
-            if (   meCoordsInt[1] - enemyCoordsInt[1] > -1
-                && meCoordsInt[1] - enemyCoordsInt[1] <= maxYincr.getValue()) {
-                // Because we have 2 option for the place (1 block higher or same h)
-                // We have to iterete between these 2 options. The first one, is the higher
-                for(int startH = 1; startH >= 0; startH--) {
-                    // If there was a no before or we found nothing
-                    if (addedStructure.to_place == null) {
-                        /// If we are above the enemy, we have to create al the structure.
-                        // How much we are going above
-                        int incr = 0;
-                        // Blocks we are going to add
-                        List<Vec3d> highSup = new ArrayList<Vec3d>();
-                        // Create the structure until it's on us
-                        while (meCoordsInt[1] > enemyCoordsInt[1] + incr) {
-                            incr++;
-                            for (int[] cordSupport : disp_surblock)
-                                highSup.add(new Vec3d(cordSupport[0], incr, cordSupport[2]));
-                        }
-                        // We go 1 block above (+1) and startH (if we looking 1 block above or not)
-                        incr += startH;
-                        // After, we are going to do a foreach. But we need also an index, so here
-                        int i = -1;
-
-                        // Lets iterate for every 4 positions
-                        for(Double[] cord_b : sur_block) {
-                            /// Note:
-                            /*
-                                Abs = Absolute
-                                Rel = Relative
-                             */
-                            /*
-                                Since they are a lot of if, i prefer keeping them
-                                separated but, also, on the same tab.
-                                I'll use "continue"
-                             */
-                            i++;
-                            /// Crystal Coordinates ///
-                            // Init + Get
-                            double[] crystalCordsAbs = {cord_b[0], cord_b[1] + incr, cord_b[2]};
-                            int[] crystalCordsRel = {disp_surblock[i][0], disp_surblock[i][1] + incr, disp_surblock[i][2]};
-
-                            /// Crystal Position Checks ///
-                            // Check, first of all, the distance
-                            if (!((distanceNowCrystal = mc.player.getDistance(crystalCordsAbs[0], crystalCordsAbs[1], crystalCordsAbs[2])) < addedStructure.distance))
-                                continue;
-                            // Check if there is enough space
-                            if  ((!((get_block(crystalCordsAbs[0], crystalCordsAbs[1], crystalCordsAbs[2])) instanceof BlockAir)
-                                || !((get_block(crystalCordsAbs[0], crystalCordsAbs[1] + 1, crystalCordsAbs[2])) instanceof BlockAir)))
-                                continue;
-                            // Check if someone is in that block
-                            if (someoneInCoords(crystalCordsAbs[0], crystalCordsAbs[2]))
-                                continue;
-                            /// Piston Coordinates ///
-                            // Init
-                            double[] pistonCordAbs = new double[3];
-                            int[] pistonCordRel = new int[3];
-                            Block tempBlock;
-
-                            /// Piston Position ///
-                            // If rotation or not betterplacement
-                            if (rotate.getValue() || !betterPlacement.getValue()) {
-                                pistonCordAbs = new double[] { crystalCordsAbs[0] + disp_surblock[i][0], crystalCordsAbs[1], crystalCordsAbs[2] + disp_surblock[i][2] };
-                                /// If statements ///
-                                // If it's not air or piston and if someone is here
-                                if ((tempBlock = get_block(pistonCordAbs[0], pistonCordAbs[1], pistonCordAbs[2])) instanceof BlockPistonBase
-                                     == tempBlock instanceof BlockAir
-                                    || someoneInCoords(pistonCordAbs[0], pistonCordAbs[2]))
-                                    // Exit
-                                    continue;
-                                // Get relative coords
-                                pistonCordRel = new int[] { crystalCordsRel[0] * 2, crystalCordsRel[1], crystalCordsRel[2] * 2};
-
-                            } else {
-                                /// Try to find the best for distance
-                                // Min distance found
-                                double distancePist = Double.MAX_VALUE;
-                                double distanceNowPiston;
-                                // Iterate for every 4 blocks
-                                for (int[] disp : disp_surblock) {
-                                    // Get Position
-                                    BlockPos blockPiston = new BlockPos(crystalCordsAbs[0] + disp[0], crystalCordsAbs[1], crystalCordsAbs[2] + disp[2]);
-                                    /// If statements ////
-                                    // If distance is not ok ok
-                                    if ((distanceNowPiston = mc.player.getDistanceSqToCenter(blockPiston)) > distancePist)
-                                        continue;
-                                    // If it's not air or piston and if someone is here
-                                    if (!(     get_block(blockPiston.x, blockPiston.y, blockPiston.z) instanceof BlockPistonBase
-                                            || get_block(blockPiston.x, blockPiston.y, blockPiston.z) instanceof BlockAir)
-                                            || someoneInCoords(crystalCordsAbs[0] + disp[0], crystalCordsAbs[2] + disp[2]))
-                                        continue;
-                                    // The block in front of the piston should be air
-                                    if (!(get_block(blockPiston.x - crystalCordsRel[0], blockPiston.y, blockPiston.z - crystalCordsRel[2]) instanceof BlockAir))
-                                        continue;
-                                    // Add new coordinates
-                                    distancePist = distanceNowPiston;
-                                    pistonCordAbs = new double[] { crystalCordsAbs[0] + disp[0], crystalCordsAbs[1], crystalCordsAbs[2] + disp[2] };
-                                    pistonCordRel = new int[] { crystalCordsRel[0] + disp[0], crystalCordsRel[1], crystalCordsRel[2] + disp[2]};
-                                }
-                                // If it change nothing
-                                if (distancePist == Double.MAX_VALUE)
-                                    continue;
-                               }
-
-                            // Checks in case you have rotate on
-                            if (rotate.getValue()) {
-                                int[] pistonCordInt = new int[] {(int) pistonCordAbs[0], (int) pistonCordAbs[1], (int) pistonCordAbs[2]};
-                                /*
-                                    We cannot allow these options:
-                                    If we are behind the piston
-                                    If we are in a vertex of a square with center the enemy
-                                    If we are 3 blocks away and not in the same x/z
-                                 */
-                                // If we are behind
-                                boolean behindBol = false;
-                                for(int checkBehind : new int[] {0,2})
-                                    // If we have an axis that is the same
-                                    if (meCoordsInt[checkBehind] == pistonCordInt[checkBehind]) {
-                                        int idx = checkBehind == 2 ? 0 : 2;
-                                        // If we are behind
-                                        if (pistonCordInt[idx] >= enemyCoordsInt[idx] == meCoordsInt[idx] >= enemyCoordsInt[idx])
-                                            behindBol = true;
-                                    }
-                                // If we are in a quarter
-                                if (!behindBol && Math.abs(meCoordsInt[0] - enemyCoordsInt[0]) == 2 && Math.abs(meCoordsInt[2] - enemyCoordsInt[2]) == 2) {
-                                    // If one of the two coords are the same and the distance is >= 2
-                                    if ( (meCoordsInt[0] == pistonCordInt[0] && (Math.abs(meCoordsInt[2] - pistonCordInt[2]) >= 2)
-                                        || (meCoordsInt[2] == pistonCordInt[2] && (Math.abs(meCoordsInt[0] - pistonCordInt[0]) >= 2))) )
-                                        behindBol = true;
-                                }
-                                // If our distance is more then 3 blocks (and one coordinate is different), exit
-                                if (!behindBol && (Math.abs(meCoordsInt[0] - enemyCoordsInt[0]) > 2 && meCoordsInt[2] != enemyCoordsInt[2])
-                                    || (Math.abs(meCoordsInt[2] - enemyCoordsInt[2]) > 2 && meCoordsInt[0] != enemyCoordsInt[0]))
-                                    behindBol = true;
-                                // Exit
-                                if (behindBol)
-                                    continue;
-
-                            }
-
-                            /// Redstone Coordinates
-                            double[] redstoneCoordsAbs = new double[3];
-                            int[] redstoneCoordsRel = new int[3];
-                            double minFound = Double.MAX_VALUE;
-                            double minNow = -1;
-                            boolean foundOne = true;
-
-                            // Iterate for all 4 positions
-                            for(int[] pos : disp_surblock) {
-                                // Get coordinates
-                                double[] torchCoords = new double[]{pistonCordAbs[0] + pos[0], pistonCordAbs[1], pistonCordAbs[2] + pos[2]};
-                                // If it's min of what we have now
-                                if ((minNow = mc.player.getDistance(torchCoords[0], torchCoords[1], torchCoords[2])) >= minFound)
-                                    continue;
-                                // if it's a redstone block, lets remove all the sides
-                                if (redstoneBlockMode && (pos[0] == pistonCordRel[0]))
-                                    continue;
-                                /*
-                                tempBlock = get_block(pistonCordAbs[0], pistonCordAbs[1], pistonCordAbs[2])) instanceof BlockPistonBase
-                                     == tempBlock instanceof BlockAir
-                                 */
-                                // Check if: Someone is here, if it's air, if it's the position of the crystal
-                                if (    someoneInCoords(torchCoords[0], torchCoords[2])
-                                    || !(   get_block(torchCoords[0], torchCoords[1], torchCoords[2]) instanceof BlockRedstoneTorch
-                                         || get_block(torchCoords[0], torchCoords[1], torchCoords[2]) instanceof BlockAir )
-                                    || (int) torchCoords[0] == (int) crystalCordsAbs[0] && (int) torchCoords[2] == (int) crystalCordsAbs[2]) {
-                                    continue;
-                                }
-                                // If the redstone torchs is in front of the piston
-                                boolean torchFront = false;
-                                for(int part : new int[] {0,2}) {
-                                    int contPart = part == 0 ? 2 : 0;
-                                    if ((int) torchCoords[contPart] == (int) pistonCordAbs[contPart] && (int) torchCoords[part] == enemyCoordsInt[part])
-                                        torchFront = true;
-                                }
-                                if (torchFront)
-                                    continue;
-                                redstoneCoordsAbs = new double[] {torchCoords[0], torchCoords[1], torchCoords[2]};
-                                redstoneCoordsRel = new int[] {pistonCordRel[0] + pos[0], pistonCordRel[1], pistonCordRel[2] + pos[2]};
-                                foundOne = false;
-                                minFound = minNow;
-                            }
-                            if (foundOne)
-                                continue;
-
-                            /// Create structure ///
-                            // If fast mode
-                            if (redstoneBlockMode && allowFastMode.getValue()) {
-                                /// Check if it's possible
-                                // First, lets if there is space for the redstone block
-                                if (get_block(redstoneCoordsAbs[0], redstoneCoordsAbs[1] - 1, redstoneCoordsAbs[2]) instanceof BlockAir) {
-                                    /// We can change everything
-                                    // Piston
-                                    pistonCordAbs = new double[] {redstoneCoordsAbs[0], redstoneCoordsAbs[1], redstoneCoordsAbs[2]};
-                                    pistonCordRel = new int[] {redstoneCoordsRel[0], redstoneCoordsRel[1], redstoneCoordsRel[2]};
-                                    // Redstone block
-                                    redstoneCoordsAbs = new double[] {redstoneCoordsAbs[0], redstoneCoordsAbs[1] - 1, redstoneCoordsRel[2]};
-                                    redstoneCoordsRel = new int[] {redstoneCoordsRel[0], redstoneCoordsRel[1] - 1, redstoneCoordsRel[2]};
-                                    // Active fastMode
-                                    fastModeActive = true;
-                                }
-                            }
-
-                            /// Create the structure
-                            // Skeleton
-                            List<Vec3d> toPlaceTemp = new ArrayList<>();
-                            int supportBlock = 0;
-
-                            /// Lets check if, under the crystal, piston and redstone there is a solid block
-                            // Crystal
-                            if(get_block(crystalCordsAbs[0], crystalCordsAbs[1] - 1, crystalCordsAbs[2]) instanceof BlockAir) {
-                                toPlaceTemp.add(new Vec3d(crystalCordsRel[0], crystalCordsRel[1] - 1, crystalCordsRel[2]));
-                                supportBlock++;
-                            }
-                            // Piston
-                            if(!fastModeActive && get_block(pistonCordAbs[0], pistonCordAbs[1] - 1, pistonCordAbs[2]) instanceof BlockAir) {
-                                toPlaceTemp.add(new Vec3d(pistonCordRel[0], pistonCordRel[1] - 1, pistonCordRel[2]));
-                                supportBlock++;
-                            }
-                            // Redstone
-                            if(get_block(redstoneCoordsAbs[0], redstoneCoordsAbs[1] - 1, redstoneCoordsAbs[2]) instanceof BlockAir) {
-                                toPlaceTemp.add(new Vec3d(redstoneCoordsRel[0], redstoneCoordsRel[1] - 1, redstoneCoordsRel[2]));
-                                supportBlock++;
-                            }
-
-                            /// Add all others blocks
-                            // Piston
-                            toPlaceTemp.add(new Vec3d(pistonCordRel[0], pistonCordRel[1], pistonCordRel[2]));
-                            // Crystal
-                            toPlaceTemp.add(new Vec3d(crystalCordsRel[0], crystalCordsRel[1], crystalCordsRel[2]));
-                            // Redstone
-                            toPlaceTemp.add(new Vec3d(redstoneCoordsRel[0], redstoneCoordsRel[1], redstoneCoordsRel[2]));
-
-                            // If we are above the enemy
-                            if (incr > 1) {
-                                // Lets add everything
-                                for (int i2 = 0; i2 < highSup.size(); i2++) {
-                                    toPlaceTemp.add(0, highSup.get(i2));
-                                    supportBlock++;
-                                }
-                            }
-
-                            /// Rotation calculation
-                            float offsetX, offsetZ, offsetY;
-                            // If horrizontaly
-                            if (disp_surblock[i][0] != 0) {
-                                offsetX = rotate.getValue() ? disp_surblock[i][0] / 2f : disp_surblock[i][0];
-                                // Check which is better for distance
-                                if (rotate.getValue()) {
-                                    if (mc.player.getDistanceSq(pistonCordAbs[0], pistonCordAbs[1], pistonCordAbs[2] + 0.5) > mc.player.getDistanceSq(pistonCordAbs[0], pistonCordAbs[1], pistonCordAbs[2] - 0.5))
-                                        offsetZ = -0.5f;
-                                    else
-                                        offsetZ = 0.5f;
-                                } else offsetZ = disp_surblock[i][2];
-                                // If vertically
-                            } else {
-                                offsetZ = rotate.getValue() ? disp_surblock[i][2] / 2f : disp_surblock[i][2];
-                                // Check which is better for distance
-                                if (rotate.getValue()) {
-                                    if (mc.player.getDistanceSq(pistonCordAbs[0] + 0.5, pistonCordAbs[1], pistonCordAbs[2]) > mc.player.getDistanceSq(pistonCordAbs[0] - 0.5, pistonCordAbs[1], pistonCordAbs[2]))
-                                        offsetX = -0.5f;
-                                    else
-                                        offsetX = 0.5f;
-                                } else offsetX = disp_surblock[i][0];
-                            }
-                            /// Calculate the y offset.
-                            // If we are above, 1, if we are belove, 0
-                            offsetY = meCoordsInt[1] - enemyCoordsInt[1] == -1 ? 0 : 1;
-
-                            // Repleace the structure
-                            addedStructure.replaceValues(distanceNowCrystal, supportBlock, toPlaceTemp, -1, offsetX, offsetZ, offsetY);
-
-                            // If trapPlayer
-                            if (blockPlayer.getValue()) {
-                                // Get the values
-                                Vec3d valuesStart = addedStructure.to_place.get(addedStructure.supportBlock + 1);
-                                // Get the opposit
-                                int[] valueBegin = new int[]{(int) -valuesStart.x, (int) valuesStart.y, (int) -valuesStart.z};
-                                // Add
-                                if (!bypassObsidian.getValue() || (int) mc.player.posY == enemyCoordsInt[1]) {
-                                    addedStructure.to_place.add(0, new Vec3d(0, incr + 1, 0));
-                                    addedStructure.to_place.add(0, new Vec3d(valueBegin[0], incr + 1, valueBegin[2]));
-                                    addedStructure.to_place.add(0, new Vec3d(valueBegin[0], incr, valueBegin[2]));
-                                    addedStructure.supportBlock += 3;
-                                }else {
-                                    addedStructure.to_place.add(0, new Vec3d(0, incr, 0));
-                                    addedStructure.to_place.add(0, new Vec3d(valueBegin[0], incr, valueBegin[2]));
-                                    addedStructure.supportBlock += 2;
-                                }
-                            }
-                            toPlace = addedStructure;
-                        }
-                    }
-                }
-            }
-            // Error h non compatible
-            else
-            yUnder = true;
-
-        }
-        catch (Exception e) {
-            printChat("Fatal Error during the creation of the structure. Please, report this bug in the discor's server", true);
-            final Logger LOGGER = LogManager.getLogger("GameSense");
-            LOGGER.error("[PistonCrystal] error during the creation of the structure.");
-            if (e.getMessage() != null)
-                LOGGER.error("[PistonCrystal] error message: " + e.getClass().getName() + " " + e.getMessage());
-            else
-                LOGGER.error("[PistonCrystal] cannot find the cause");
-            int i5 = 0;
-
-            if (e.getStackTrace().length != 0) {
-                LOGGER.error("[PistonCrystal] StackTrace Start");
-                for(StackTraceElement errorMess : e.getStackTrace()) {
-                    LOGGER.error("[PistonCrystal] " + errorMess.toString());
-                }
-                LOGGER.error("[PistonCrystal] StackTrace End");
-            }
-
-            if (aimTarget != null) {
-                LOGGER.error("[PistonCrystal] closest target is not null");
-            }else LOGGER.error("[PistonCrystal] closest target is null somehow");
-            for (Double[] cord_b : sur_block) {
-                if (cord_b != null) {
-                    LOGGER.error("[PistonCrystal] " + i5 + " is not null");
-                }else {
-                    LOGGER.error("[PistonCrystal] " + i5 + " is null");
-                }
-                i5++;
-            }
-
-        }
-
-        if (debugMode.getValue() && addedStructure.to_place != null) {
-            for(Vec3d parte : addedStructure.to_place) {
-                printChat(String.format("%f %f %f", parte.x, parte.y, parte.z), false);
-            }
-        }
-
-        return addedStructure.to_place != null;
-    }
-
-    public static boolean someoneInCoords(double x, double z) {
-        int xCheck = (int) x;
-        int zCheck = (int) z;
-        // Get player's list
-        List<EntityPlayer> playerList = mc.world.playerEntities;
-        // Iterate
-        for(EntityPlayer player : playerList) {
-            // I dont think you need to check also the yLevel
-            if ((int) player.posX == xCheck && (int) player.posZ == zCheck)
-                return true;
-        }
-
-        return false;
-    }
-
-    // Get all the materials
-    private boolean getMaterialsSlot() {
-		/*
-			// I use this as a remind to which index refers to what
-			0 => obsidian
-			1 => piston
-			2 => Crystals
-			3 => redstone
-			4 => sword
-			5 => pick
-		 */
-
-        if (mc.player.getHeldItemOffhand().getItem() instanceof ItemEndCrystal) {
-            slot_mat[2] = 11;
-        }
-
-        // Iterate for all the inventory
-        for(int i = 0; i < 9; i++) {
-            ItemStack stack = mc.player.inventory.getStackInSlot(i);
-
-            // If there is no block
-            if (stack == ItemStack.EMPTY){
-                continue;
-            }
-            // If endCrystal
-            if (slot_mat[2] == -1 && stack.getItem() instanceof ItemEndCrystal) {
-                slot_mat[2] = i;
-            // If sword
-            }else if (antiWeakness.getValue() && stack.getItem() instanceof ItemSword) {
-                slot_mat[4] = i;
-            }else
-            // If Pick
-            if (!placeMode.getValue().equals("Torch") && stack.getItem() instanceof ItemPickaxe) {
-                slot_mat[5] = i;
-            }
-            if (stack.getItem() instanceof ItemBlock){
-
-                // If yes, get the block
-                Block block = ((ItemBlock) stack.getItem()).getBlock();
-
-                // Obsidian
-                if (block instanceof BlockObsidian) {
-                    slot_mat[0] = i;
-                } else
-                    // PistonBlock
-                    if (block instanceof BlockPistonBase) {
-                        slot_mat[1] = i;
-                    } else
-                        // RedstoneTorch / RedstoneBlock
-                        if (!placeMode.getValue().equals("Block") && block instanceof BlockRedstoneTorch) {
-                            slot_mat[3] = i;
-                            redstoneBlockMode = false;
-                        }
-                        else if (!placeMode.getValue().equals("Torch") &&block.translationKey.equals("blockRedstone")) {
-                            slot_mat[3] = i;
-                            redstoneBlockMode = true;
-                        }
-            }
-        }
-        // Count what we found
-        int count = 0;
-        for(int val : slot_mat) {
-            if (val != -1)
-                count++;
-        }
-
-        // If we have everything we need, return true
-        return count == 4 + (antiWeakness.getValue() ? 1 : 0) + (redstoneBlockMode ? 1 : 0);
-
-    }
-
-    // Check if it's in a hole
-    private boolean is_in_hole() {
-        sur_block = new Double[][] {
-                {aimTarget.posX + 1, aimTarget.posY, aimTarget.posZ},
-                {aimTarget.posX - 1, aimTarget.posY, aimTarget.posZ},
-                {aimTarget.posX, aimTarget.posY, aimTarget.posZ + 1},
-                {aimTarget.posX, aimTarget.posY, aimTarget.posZ - 1}
-        };
-
-        
-        // Check if the guy is in a hole
-        return !(get_block(sur_block[0][0], sur_block[0][1], sur_block[0][2]) instanceof BlockAir) &&
-                !(get_block(sur_block[1][0], sur_block[1][1], sur_block[1][2]) instanceof BlockAir) &&
-                !(get_block(sur_block[2][0], sur_block[2][1], sur_block[2][2]) instanceof BlockAir) &&
-                !(get_block(sur_block[3][0], sur_block[3][1], sur_block[3][2]) instanceof BlockAir);
-    }
-
-    // Get block from coordinates
-    private Block get_block(double x, double y, double z) {
-        return mc.world.getBlockState(new BlockPos(x, y, z)).getBlock();
-    }
-
-    // Find player you are looking
-    public static EntityPlayer findLookingPlayer(double rangeMax) {
-        // Get player
-        ArrayList<EntityPlayer> listPlayer = new ArrayList<>();
-        // Only who is in a distance of enemyRange
-        for(EntityPlayer playerSin : mc.world.playerEntities) {
-            if (basicChecksEntity(playerSin))
-                continue;
-            if (mc.player.getDistance(playerSin) <= rangeMax) {
-                listPlayer.add(playerSin);
-            }
-        }
-
-        EntityPlayer target = null;
-        // Get coordinate eyes + rotation
-        Vec3d positionEyes = mc.player.getPositionEyes(mc.getRenderPartialTicks());
-        Vec3d rotationEyes = mc.player.getLook(mc.getRenderPartialTicks());
-        // Precision
-        int precision = 2;
-        // Iterate for every blocks
-        for(int i = 0; i < (int) rangeMax; i++) {
-            // Iterate for the precision
-            for(int j = precision; j > 0 ; j--) {
-                // Iterate for all players
-                for(EntityPlayer targetTemp : listPlayer) {
-                    // Get box of the player
-                    AxisAlignedBB playerBox = targetTemp.getEntityBoundingBox();
-                    // Get coordinate of the vec3d
-                    double xArray = positionEyes.x + (rotationEyes.x * i) + rotationEyes.x/j;
-                    double yArray = positionEyes.y + (rotationEyes.y * i) + rotationEyes.y/j;
-                    double zArray = positionEyes.z + (rotationEyes.z * i) + rotationEyes.z/j;
-                    // If it's inside
-                    if (   playerBox.maxY >= yArray && playerBox.minY <= yArray
-                            && playerBox.maxX >= xArray && playerBox.minX <= xArray
-                            && playerBox.maxZ >= zArray && playerBox.minZ <= zArray) {
-                        // Get target
-                        target = targetTemp;
-                    }
-                }
-            }
-        }
-
-        return target;
-    }
-
-    // Basic checks for an entity
-    public static boolean basicChecksEntity(Entity pl) {
-        return pl == mc.player || Friends.isFriend(pl.getName()) || pl.isDead;
-    }
-
-    // Find closest target
-    public static EntityPlayer findClosestTarget(double rangeMax, EntityPlayer aimTarget){
-        List<EntityPlayer> playerList = mc.world.playerEntities;
-
-        EntityPlayer closestTarget_test = null;
-
-        for (EntityPlayer entityPlayer : playerList){
-
-            if (basicChecksEntity(entityPlayer))
-                continue;
-
-            if (aimTarget == null && mc.player.getDistance(entityPlayer) <= rangeMax){
-                closestTarget_test = entityPlayer;
-                continue;
-            }
-            if (aimTarget != null && mc.player.getDistance(entityPlayer) <= rangeMax && mc.player.getDistance(entityPlayer) < mc.player.getDistance(aimTarget)){
-                closestTarget_test = entityPlayer;
-            }
-        }
-        return closestTarget_test;
-    }
-
-    // PrintChat
-    private void printChat(String text, Boolean error) {
-        MessageBus.sendClientPrefixMessage((error ? ColorMain.getDisabledColor() : ColorMain.getEnabledColor()) + text);
-    }
-
-    /// AutoCrystal break things ///
-    public static void lookAtPacket(double px, double py, double pz, EntityPlayer me) {
-        double[] v = calculateLookAt(px, py, pz, me);
-        setYawAndPitch((float) v[0], (float) v[1]);
-    }
-    public static double[] calculateLookAt(double px, double py, double pz, EntityPlayer me) {
-        double dirx = me.posX - px;
-        double diry = me.posY - py;
-        double dirz = me.posZ - pz;
-
-        double len = Math.sqrt(dirx*dirx + diry*diry + dirz*dirz);
-
-        dirx /= len;
-        diry /= len;
-        dirz /= len;
-
-        double pitch = Math.asin(diry);
-        double yaw = Math.atan2(dirz, dirx);
-
-        pitch = pitch * 180.0d / Math.PI;
-        yaw = yaw * 180.0d / Math.PI;
-
-        yaw += 90f;
-
-        return new double[]{yaw,pitch};
-    }
-    private static boolean isSpoofingAngles;
-    private static double yaw;
-    private static double pitch;
-    public static void setYawAndPitch(float yaw1, float pitch1) {
-        yaw = yaw1;
-        pitch = pitch1;
-        isSpoofingAngles = true;
-    }
-    public static void breakCrystal(Entity crystal) {
-        mc.playerController.attackEntity(mc.player, crystal);
-        mc.player.swingArm(EnumHand.MAIN_HAND);
-    }
-    @EventHandler
-    private final Listener<PacketEvent.Send> packetSendListener = new Listener<>(event -> {
-        Packet packet = event.getPacket();
-        if (packet instanceof CPacketPlayer) {
-            if (isSpoofingAngles) {
-                ((CPacketPlayer) packet).yaw = (float) yaw;
-                ((CPacketPlayer) packet).pitch = (float) pitch;
-            }
-        }
-    });
-    public static void resetRotation() {
-        if (isSpoofingAngles) {
-            yaw = mc.player.rotationYaw;
-            pitch = mc.player.rotationPitch;
-            isSpoofingAngles = false;
-        }
-    }
-
-}
